update instance method test to use another file for subclass
