+++ tests/instance_method_subclass_cjs/mod.js
@@ -1,13 +1,3 @@
-const assert = require('node:assert')
-
 class Base {}
 
-class Undici extends Base {
-  async fetch (url, tag) {
-    assert.strictEqual(url, 'https://example.com')
-    assert.strictEqual(tag, 'mutated')
-    return 42
-  }
-}
-
-module.exports = { Undici }
+module.exports = { Base }
