+++ tests/instance_method_subclass_cjs/subclass.js
@@ -1,0 +1,12 @@
+const assert = require('node:assert')
+const { Base } = require('./instrumented.js')
+
+class Undici extends Base {
+  async fetch (url, tag) {
+    assert.strictEqual(url, 'https://example.com')
+    assert.strictEqual(tag, 'mutated')
+    return 42
+  }
+}
+
+module.exports = { Undici }
